requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
PyPDF2>=3.0.0
reportlab>=4.0.0
uuid
json5
tempfile
datetime